Resolves #4290 - no clear documentation on swap input plugin
